Refactor to avoid having to needlessly set BEAMLINE when running murfey.hub
