+++ src/murfey/cli/add_user.py
@@ -4,8 +4,8 @@ import os
 from sqlmodel import Session, create_engine
 
 from murfey.server.auth import hash_password
-from murfey.server.config import get_machine_config
 from murfey.server.murfey_db import url
+from murfey.util.config import get_machine_config
 from murfey.util.db import MurfeyUser as User
