+++ src/murfey/cli/db_sessions.py
@@ -2,8 +2,8 @@ import argparse
 
 from sqlmodel import Session, create_engine, select
 
-from murfey.server.config import get_machine_config
 from murfey.server.murfey_db import url
+from murfey.util.config import get_machine_config
 from murfey.util.db import Session as MurfeySession
