+++ src/murfey/cli/inject_spa_processing.py
@@ -8,9 +8,9 @@ import workflows
 import zocalo
 from sqlmodel import Session, create_engine, select
 
-from murfey.server.config import get_machine_config, get_microscope
 from murfey.server.ispyb import TransportManager
 from murfey.server.murfey_db import url
+from murfey.util.config import get_machine_config, get_microscope
 from murfey.util.db import (
     AutoProcProgram,
     ClientEnvironment,
