+++ src/murfey/server/auth/__init__.py
@@ -8,8 +8,8 @@ from jose import JWTError, jwt
 from passlib.context import CryptContext
 from sqlmodel import Session, create_engine, select
 
-from murfey.server.config import get_machine_config
 from murfey.server.murfey_db import url
+from murfey.util.config import get_machine_config
 from murfey.util.db import MurfeyUser as User
 from murfey.util.db import Session as MurfeySession
