+++ src/murfey/server/demo_api.py
@@ -37,8 +37,8 @@ from murfey.server import (
 from murfey.server import shutdown as _shutdown
 from murfey.server import templates
 from murfey.server.auth import validate_token
-from murfey.server.config import MachineConfig, from_file
 from murfey.server.murfey_db import murfey_db
+from murfey.util.config import MachineConfig, from_file
 from murfey.util.db import (
     AutoProcProgram,
     ClientEnvironment,
